requirements.txt
absl-py==0.2.0
arff2pandas==1.0.1
asn1crypto==0.24.0
astor==0.6.2
bleach==1.5.0
catboost==0.8.1
certifi==2018.4.16
cffi==1.11.4
chardet==3.0.4
click==6.7
conda==4.4.10
cryptography==2.1.4
cycler==0.10.0
Cython==0.28.2
docutils==0.14
Flask==1.0.2
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imbalanced-learn==0.3.3
imblearn==0.0
itsdangerous==0.24
javabridge==1.0.17
Jinja2==2.10
Keras==2.1.6
kiwisolver==1.0.1
liac-arff==2.2.1
lightgbm==2.1.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
menuinst==1.4.11
mkl-fft==1.0.0
mkl-random==1.0.1
mlxtend==0.12.0
numpy==1.14.2
pandas==0.22.0
protobuf==3.5.2.post1
pycosat==0.6.3
pycparser==2.18
pymongo==3.6.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pyswarm==0.6
python-dateutil==2.7.2
python-weka-wrapper3==0.1.4
pytz==2018.4
pywin32==222
PyYAML==3.12
requests==2.18.4
rpy2==2.9.3
ruamel-yaml==0.15.35
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
tzlocal==1.5.1
urllib3==1.22
Werkzeug==0.14.1
win-inet-pton==1.0.1
wincertstore==0.2
xgboost==0.71
